breaking the homeVIew into BPI arc
